Treat char typed data like int or byte rather than like string (bug 114)

+++ core/src/org/cougaar/core/util/XMLPrinter.java
@@ -90,7 +90,7 @@ public class XMLPrinter extends AsciiPrinter {
     print("<");
     print(name);
     print(">");
-    print(c);
+    print((int)c);
     print("</");
     print(name);
     print(">");
